requirements.txt
apscheduler
crewai
markdown2
ollama
pillow
pystray
python-dotenv
requests
youtube-transcript-api
